improved some code and tests
