--- src/authn/IdentityProviders.ts
@@ -1,19 +1,0 @@
-const DEFAULT_ISSUERS = [
-    {
-      name: 'Solid Community',
-      uri: 'https://solidcommunity.net'
-    },
-    {
-      name: 'Solid Web',
-      uri: 'https://solidweb.org'
-    },
-    {
-      name: 'Inrupt.net',
-      uri: 'https://inrupt.net'
-    },
-    {
-      name: 'pod.Inrupt.com',
-      uri: 'https://broker.pod.inrupt.com'
-    }
-  ]
-
