+++ src/util/UtilityLogic.ts
@@ -1,7 +1,5 @@
-import { NamedNode, Node, st, term, sym, Statement } from "rdflib";
+import { NamedNode, Statement, sym } from "rdflib";
 import { LiveStore, SolidNamespace } from "../index";
-import { ProfileLogic } from "../profile/ProfileLogic";
-import { newThing } from "./uri";
 
 export const ACL_LINK = sym(
   "http://www.iana.org/assignments/link-relations/acl"
